pyproject.toml
[tool.poetry]
name = "attract-ia-tcc"
version = "0.1.0"
description = ""
authors = ["Matheus <matheusgalvain@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.3.3"
flask-session = "^0.5.0"
mysql-connector-python = "^8.1.0"
flake8 = "^6.1.0"
openai = "^0.28.0"
requests = "^2.31.0"
pytest = "^7.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
